+++ project-spring/src/main/java/toyproject/todoList/domain/comment/service/dto/UpdateCommentRequest.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class UpdateCommentRequest {
 
-    private Long commentId;
+    private Integer commentId;
 
     private String content;
 }
