+++ project-spring/src/main/java/toyproject/todoList/domain/recomment/repository/ReCommentRepository.java
@@ -3,5 +3,5 @@ package toyproject.todoList.domain.recomment.repository;
 import toyproject.todoList.domain.recomment.entity.ReComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReCommentRepository extends JpaRepository<ReComment, Long> {
+public interface ReCommentRepository extends JpaRepository<ReComment, Integer> {
 }
