+++ project-spring/src/main/java/toyproject/todoList/domain/recomment/service/dto/UpdateReCommentRequest.java
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class UpdateReCommentRequest {
-    private Long reCommentId;
+    private Integer reCommentId;
     private String content;
 
 }
