+++ project-spring/src/main/java/toyproject/todoList/domain/tag/TagRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TagRepository extends JpaRepository<Tag, Long> {
+public interface TagRepository extends JpaRepository<Tag, Integer> {
     List<Tag> findByPost(Post post);
 }
