+++ project-spring/src/main/java/toyproject/todoList/domain/todo/repository/TodoRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface TodoRepository extends JpaRepository<TodoList, Long>, TodoCustomRepository {
+public interface TodoRepository extends JpaRepository<TodoList, Integer>, TodoCustomRepository {
 
 }
